+++ lib/statix/conn.ex
@@ -8,13 +8,21 @@ defmodule Statix.Conn do
   require Logger
 
   def new(host, port) when is_binary(host) do
-    new(string_to_charlist(host), port)
+    new(String.to_charlist(host), port)
   end
 
   def new(host, port) when is_list(host) or is_tuple(host) do
-    {:ok, addr} = :inet.getaddr(host, :inet)
-    header = Packet.header(addr, port)
-    %__MODULE__{header: header}
+    case :inet.getaddr(host, :inet) do
+      {:ok, address} ->
+        header = Packet.header(address, port)
+        %__MODULE__{header: header}
+
+      {:error, reason} ->
+        raise(
+          "cannot get the IP address for the provided host " <>
+            "due to reason: #{:inet.format_error(reason)}"
+        )
+    end
   end
 
   def open(%__MODULE__{} = conn) do
@@ -51,11 +59,5 @@ defmodule Statix.Conn do
           {:inet_reply, _port, status} -> status
         end
     end
-  end
-
-  if Version.match?(System.version(), ">= 1.3.0") do
-    defp string_to_charlist(string), do: String.to_charlist(string)
-  else
-    defp string_to_charlist(string), do: String.to_char_list(string)
   end
 end